F VICTOIRE

+++ main.c
@@ -7,13 +7,13 @@
 
 int main(int argc, const char *argv[])
 {
-	long double		bite;
+	double		bite;
 
 //	bite = 4.94065645841246544176568792868E-324;
 	bite = 3.14;
-	ft_printf("%Lf\n", bite);
+	ft_printf("%f\n", bite);
 //	printf("\n");
-	printf("%Lf\n", bite);
+	printf("%.52f\n", bite);
 	//%      i 42 == | 42|
 	//	% i -42 == |-42|
 	//	% 4i 42 == |  42|
